Reject messages instead of requeue when failing to write
Can alternatively set a TTL for a message before requeuing or pushing
it.  That can be implemented in the future (if needed) as might not be
trivial and might require to push a new message to the exchange, since
is not clear wether TTL can be set on a Message directly before
requeuing it.